# backend/static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "API para gerenciamento de tarefas para estudantes",
      "version": "1.0.0",
      "title": "Nota10 API"
    },
    "host": "127.0.0.1:5000",
    "basePath": "/",
    "tags": [
      {
        "name": "tarefas",
        "description": "Operações relacionadas a tarefas"
      }
    ],
    "paths": {
      "/cadastrar_tarefa": {
        "post": {
          "tags": ["tarefas"],
          "summary": "Cadastrar uma nova tarefa",
          "description": "Esta rota cadastra uma nova tarefa.",
          "operationId": "cadastrarTarefa",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Tarefa a ser cadastrada",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Tarefa"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Tarefa cadastrada com sucesso",
              "schema": {
                "$ref": "#/definitions/Tarefa"
              }
            }
          }
        }
      },
      "/buscar_tarefas": {
        "get": {
          "tags": ["tarefas"],
          "summary": "Buscar todas as tarefas",
          "description": "Esta rota busca todas as tarefas cadastradas.",
          "operationId": "buscarTarefas",
          "produces": ["application/json"],
          "responses": {
            "200": {
              "description": "Lista de tarefas",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Tarefa"
                }
              }
            }
          }
        }
      },
      "/editar_tarefa/{id}": {
        "put": {
          "tags": ["tarefas"],
          "summary": "Editar uma tarefa existente",
          "description": "Esta rota edita uma tarefa existente com base no ID fornecido.",
          "operationId": "editarTarefa",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID da tarefa a ser editada",
              "required": true,
              "type": "integer",
              "format": "int64"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Tarefa com dados atualizados",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Tarefa"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Tarefa editada com sucesso",
              "schema": {
                "$ref": "#/definitions/Tarefa"
              }
            }
          }
        }
      },
      "/deletar_tarefa/{id}": {
        "delete": {
          "tags": ["tarefas"],
          "summary": "Deletar uma tarefa existente",
          "description": "Esta rota deleta uma tarefa existente com base no ID fornecido.",
          "operationId": "deletarTarefa",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID da tarefa a ser deletada",
              "required": true,
              "type": "integer",
              "format": "int64"
            }
          ],
          "responses": {
            "200": {
              "description": "Tarefa deletada com sucesso",
              "schema": {
                "$ref": "#/definitions/Tarefa"
              }
            }
          }
        }
      }
    },
    "definitions": {
      "Tarefa": {
        "type": "object",
        "required": ["titulo", "prazo"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "prazo": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
  